pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gita-autoposter"
version = "0.1.0"
description = "Agentic workflow scaffold for Bhagavad Gita auto posting."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.5",
  "python-dotenv>=1.0",
  "Pillow>=10.0",
  "openpyxl>=3.1",
  "pandas>=2.0",
  "requests>=2.31",
  "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
gita-autoposter = "gita_autoposter.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
